requirements.txt
django >4.2,<4.3
python-decouple
djangorestframework
drf-yasg[validation]
django-mail-templated
cryptography